remove dead code; rename new methods
